Update Firebase configuration
